doc(design): focus on Kubernetes
Changes the design to focus on Kubernetes only. While it is nice to have other
systems around as well, the problems solved by tanka really only occur in the
context of Kubernetes.

This means tanka should focus on making the Kubernetes experience as good as
possible instead of complicating the overall system by the presence of providers