feat(cmd): allow the baseDir to be passed as an argument (#6)
The baseDir is the most important directory, because it directly affects JPATH
resolution.

It was assumed as the pwd so far. However, it is handy in scripts to specify it
on the command line. This behaviour switches to specify only.

BREAKING: You must specify the baseDir on the command line. To mimic the old
behaviour, use `tk show .`