feat(cli): environment manipulation (#73)
Adds a set of commands to work with and manipulate environments:

- `tk env add`: Create a new environment, optionally setting parameters (now also used by `tk init`)- 
- `tk env remove` (`tk env rm`): Remove an environment
- `tk env list` (`tk env ls`): List environments (optionally as JSON)
- `tk env set`: set properties on environments (the same available for tk env add)

These commands aim to be `ks env` compatible. However, Tanka infers the environments name from the directory, so it is not possible to modify it using `add` / `set --name`. Just `mv` the directory in that case (included in the error message).