feat(jsonnet): tk standard lib draft
Adds a first draft of the extended tk standard library.

This library is meant as a soft wrapper around std.native() calls
special to Tanka only, to decouple the internals and the user facing API.

This especially allows to change the underlying native func as needed,
while keeping the API consistent at the jsonnet level.