Support lists when building k8s manifests
This commit modifies the generation of kubernetes manifests so that we
can support lists of resources. As an example, where before the
following jsonnet would fail:

    ```json
    {
      resources: [
        {
          apiVersion: "v1",
          kind: "ConfigMap",
          metadata: {
            name: "config",
          },
        }
      ],
    }
    ```

We will now return a valid configmap resource.

In addition to this change, we are also stricter on how we detect
kubernetes resources. Where before we checked for kind and
metadata.name, we now also require apiVersion, as is required by
kubernetes API conventions.

Whilst making these changes, we've also tidied up the tests for building
manifests. This included extracting common fixtures into a testdata
directory, where they can be represented as much more readable json. We
have also renamed the `Reconcile` function to be `Compile`, as the term
reconcile is overloaded in the kubernetes context (think controller
reconcile loops) and have simplified the filtering logic.