refactor(kubernetes): client interface; feat(cli): status (#111)
* refactor(kubernetes): client interface

* feat(cli): status command

Adds a `tk status <path>` command to quickly view information about a certain
environment, including objectspecs, env and context

* refactor(kubernetes): make Reconcile static

The reconcile function does not rely on any `Kubernetes` specific things, so it
makes no sense to have it included in there.

This also simplifies the use of Reconciling, as creating a Kubernetes requires a
correctly set up client.

* fix(kubernetes): do not set empty namespace

* refactor(kubernetes): handle context in constructor

Handles the context discovery once in the constructor instead of on every single
call.

This was not possible before because we had actions that had to run without a
context on the Kubernetes object. These are moved elsewhere, so we can iron this
out!

* chore: goimports

* fix(kubernetes): adapt subsetdiff to refactor

Wasn't properly adapted before, wasn't actually working at all

* fix(kubernetes): purposely panic on yaml errors

It should never happen that the yaml marshal fails. If it does, it's most
probably a programming issue that needs to be addressed, rather than the error
handled.

* feat(kubernetes): automatic manifest verification

Manifests are now always verified a very basic schema

* feat(kubernetes): maintain track of the path at all times

* test(kubernetes): remove map-order related flake

* fix(cli): improve status description

* fix: golint, review feedback