fix: stable naming of the environment
Before, the name of the internally used Environment type was different depending
on the argument passed on the command line, even though it resulted in the same
environment after all (relative vs absolute paths, etc.)

This fixes the issue by using the `jpath` package to compute baseDir and
rootDir, using the relative path from rootDir to baseDir as the name. This
results in stable names, regardless which directory representation is used.