Fix handling of indeterminate depth of arrays
This change brings in the ``reflect`` module, which appears to be
needed to supplement the lack of generics, or advanced handling in
Go's type switches.

This is unfortunate as the functionality inside of
the ``reflect`` modules often slows down the runtime considerably,
however a similar technique is used inside of ``kubectl``, among other
CLIs. It should probably be avoided if attempting to serve lots of
clients in an available web API, however the slowdown should be
relatively minor in typical usage of tanka.

Great effort has been taken to keep the changes in the code to as low
a level as possible, whilst retaining as familiar structure as
possible.