refactor(kubernetes): .jsonnet test fixtures
Moves the testdata for `pkg/kubernetes` out of Go, into several
`.jsonnet` files for easier composing.

This also ensures that the data passed to `Extract()` and `Reconcile()`
is exactly what would be used in real world. This is especially
important because Go is usually typed, but `json.Unmarshal` returns
untyped `interface{}` objects.