fix: Stable environment name (#182)
* fix: stable naming of the environment

Before, the name of the internally used Environment type was different depending
on the argument passed on the command line, even though it resulted in the same
environment after all (relative vs absolute paths, etc.)

This fixes the issue by using the `jpath` package to compute baseDir and
rootDir, using the relative path from rootDir to baseDir as the name. This
results in stable names, regardless which directory representation is used.

* refactor(api): simplify

* fix(cli): remove blank line from output

* fix(cli/jsonnet): do static import analysis on environment level

All of our current commands take environments, only the import analysis works on
plain files ... this has no benefit, it actually makes the implementation
harder, as we cannot use stable paths there.

Switches to environments as the target.