refactor(kubernetes): Jsonnet testdata (#181)
* refactor(kubernetes): .jsonnet test fixtures

Moves the testdata for `pkg/kubernetes` out of Go, into several
`.jsonnet` files for easier composing.

This also ensures that the data passed to `Extract()` and `Reconcile()`
is exactly what would be used in real world. This is especially
important because Go is usually typed, but `json.Unmarshal` returns
untyped `interface{}` objects.

* fix(kubernetes): extract without reflect

In a recent change (3903b9a86a27b9c570d4433b382e3b93be193745), we had
the impression that it was required to handle nested slices separately,
because there did not appear to be a `[]interface{}`.

As it turns out, we tricked ourselves a bit: While Go does indeed
differentiate between these as observed, `json.Unmarshal` creates only
`[]interface{}`, regardless of what lies underneath.

Because our testdata was written in Go instead of json(net), we assumed
a complexity that won't ever happen in real life, because we only deal
with JSON after all.

This implementation now is much more straightforward and faster, because
we don't need `reflect` anymore!