cmd/tk: fix paging issues with colored diff output
The less command has two flags: --raw-control-chars and
--RAW-CONTROL-CHARS, each of which keeps ASCII control characters
present in the paginated output. The former flag was being used
previously, and the man page warns about various display problems
occurring when using that flag.

On the other hand, --RAW-CONTROL-CHARS only keeps color escape
sequences, and the man page notes that appearance will be maintained
correctly in most cases. In my local testing, this appears to be true
and colorized output in less will now work as expected.

Fixes #82.