chore(packaging): windows/amd64 binary
Adds `windows/amd64` platform, so we also provide `.exe` files for users
interested in Tanka on native Windows.

Everybody please keep in mind this is community support only, Grafana
Labs will not actively develop on Windows and not actively make Tanka
compatible with that platform.

Yet we try not to break it and accept community patches for arising issues.