doc: refresh website (#285)
* doc: refresh website

Adds the new Logo to the website and also rebuilds the landing page to
make use of some advanced layouting React gives us.

Still lot to do on the website, but this gets us started

* feat(docs): /install tabs

Adds a new component to the /install route, to show tabs for each
supported platform / package manager

* style(docs): s/obviously/clearly

* fix(docs): jb install links

* feat(docs): algolia search

Adds full page text search using Algolia DocSearch, which is kindly
provided to us for free as we are an open-source project.

* fix(docs): algolia ssr