feat(process): inject metadata.namespace
Based on information from the `resources` package, we can now reliably
tell whether resources are namespaced or not. Using that info, we can
add namespaces to resources that need this during `show` and `export`.

For CRD's, we naturally can't know this in advance. For now, the
annotation `tanka.dev/namespaced` can be set to `"true"` (`string`, not
`bool`), to enable namespace injection.

For the future, we could provide a mechanism to dynamically ask the
cluster if it is available.