feat: Delete Resources (#313)
* Implement `tk delete` functionality.

This can be used to tear down a tanka environment. When doing `tk
delete`, tanka will generate all manifests for an environment and remove
them from kubernetes cluster.

This is a safer variant of the one-liner mentioned at https://github.com/grafana/tanka/issues/155

Signed-off-by: Milan Plzik <milan.plzik@grafana.com>

* Move k8s object reversing to kube.Delete.

In order to avoid cascading objects deletions on server side, we need
to sort the object in an order reverse to the one used fpr Apply(). This
is best done directly in kube.Delete(), making the rest of tanka
codebase agnostic to generated object ordering.

Closes https://github.com/grafana/tanka/issues/155