feat(helmraiser): automatic repository resolution
Currently, the user is tasked with properly configuring their Helm so
that repositories are available, etc.

This is contrary to our effort of keeping the config self-contained,
etc.

Turns out, Tanka can actually do that for the user and also quite well.

The trick is to use a custom `.config/helm/repositories.yml` at a
temporary location where we set-up a single repository to pull from,
which is passed from Jsonnet. This keeps Chart and Repository in the
same place in Jsonnet, similar to how docker images have repo and name
in the same place as well.

The use of the temp file avoids tampering with their actual Helm setup.

If accepted, I'll also update helm-util to provide helpers to use this
more easily.