feat(helm): Require Charts to be locally available
This modifies the `helmTemplate` native function in a way that it
expectes a relative path instead of the `repo/name` identifier for
Charts.

Before, this feature relied on the system wide Helm cache, which is
independent from the Tanka project and not controlled by git at all.

This does not play well with out aim for hermeticity and strong
reproducibility.

To overcome this, `helmTemplate` resolves the chart **relative to the
calling file**. This suggests to keep Helm Charts as close to the
consuming source code as possible.