refactor(api): Struct API (BREAKING)
This converts all of the public API to no longer take a wild variety of
modifier functions, but instead clearly defined structs that only allow
setting those properties, that are actually relevant to that specific
struct.

This not only reduces the amount of boilerplate required, but also makes
the API more discoverable and defines it more clearly.

THIS IS A BREAKING CHANGE