fix(helm): ignore empty objects
Apparently, helm does sometimes return elements in the YAML stream that
consist only of comments – meaning they are empty.

Because they don't pass our validation code, and also because they are
useless we need to ignore them.

Also fixes the name layout, putting `metadata.name` first and `kind`
last, as it was in the original helmraiser PR