fix(cli): Environment name at env creation (#404)
Environment name is the relative path of an env dir. tk env add is incorrectly
using the base dir name. This PR fixes it.

jpath pkg is used to benefit from the existing implementation. FindRoot function
is made public instead of using Resolve because Resolve does many other things
as well, some of which require files that don't exist at that time.

Signed-off-by: Michal Wasilewski <mwasilewski@gmx.com>