feat(jsonnet): allow other entrypoints (#389)
So far, Tanka always evaluates `main.jsonnet` whatever directory it receives from the user. Going on, this is still supported and the default behavior.

When passed a plain file instead of a directory instead however, Tanka evaluates that file now, choosing the enclosing directory as the environments directory. 

This makes `tk show` and `tk eval` more lightweight, requiring only a single file, not a whole directory.