fix(tools): normalize tk fmt paths (#411)
* fmt: Normalize file paths on windows to honor exclude globs

The glob package does not treat path separators as special, so patterns
like `**/vendor/**` literally match a path containing `/vendor/`. This
breaks down on windows. We can fix this by using filepath.ToSlash(...)
to normalize the slash direction when looking for files to format.

* jsonnet: TransitiveImports: Normalize paths for windows

This fixes a failing test on windows due to path inconsistent path
separators. The test now passes on windows and linux.