fix(jpath): support nested calling again
Originally, Tanka was able to find the environments entrypoint, even
when it was actually called on a directory that is nested underneath
that environment.

This was modeled after the ability to use `git` commands anywhere in the
project.

This functionality was lost however in a recent PR.

While restoring it, I took the chance to do a long overdue refactoring
of the `jpath` package.