fix(api): Prevent race condition
Because jsonnet.Opts includes a `map[string]string`, it is not
concurrency-safe, as all goroutines effectively share the same extCode
and tlaCode.

Modifications of any routine is also reflected onto every other routine.

To temporarily combat that, this PR deep-clones jsonnet.Opts before
passing it to any goroutine.

This is not a good permanent solution, because it can be easily
forgotten. We should instead look into ways of fixing the jsonnet
package so that such race conditions don't occur at all