fix(cli): Split diff and non-diff output (#537)
On commands that produce diffs: diff, apply, and prune.

It is desirable for some use cases to parse diff output from `kubectl
diff` (e.g. secret redaction until
https://github.com/kubernetes/kubernetes/issues/87840 is resolved). This
is a lot easier when diff and non-diff output are not combined on the
same channel, stdout.

This commit moves diagnostic, non-diff output for the 3 relevant
subcommands to stderr.