fix: make export --format respect "/" in template actions (#572)
During tk export path separators are temporarily replaced with BEL in
order to protect paths from being mixed with template output. But when
there is a slash in a template action (e.g. {{index .metadata.labels
"app.kubernetes.io/name"}}, that slash also got replaced, and the
template did not work as expected.

This fix ensures that the BEL replacement only occurs within the text
portions of the template, and the action blocks are being preserved.

Fixes #568