use yaml.v3 for serialising manifests
This commit fixes the issue exacerbated by the test submitted in the
parent commit where the presence of a tab character in the input string
for the configmap's data caused the output to be scrambled.

While the true root cause is not entirely known (and is supposedly fixed
on v3 of the YAML library), the indentation of the output manifest
switches to 4 spaces instead of 2, and the tabs are handled properly,
making the manifest applicable on a kubernetes cluster.