Export Caching: Fix caching in case of missing import (#712)
Closes https://github.com/grafana/tanka/issues/683

The issue was `doc-util`. It is used but not vendored (as expected here: https://github.com/jsonnet-libs/docsonnet).
jsonnet doesn't care about paths that aren't used so the export cache shouldn't either
This PR changes the hash calculation process to ignore missing files. If a file is needed, jsonnet will catch it during the evaluation process and cause an error there

_Also documented the cache_