`getSnippetHash`: Use regexp instead of parsing whole AST (#759)
* `getSnippetHash`: Use regexp instead of parsing whole AST
When calculating the env hash to use as cache keys, we don't need an exact list of imports
Therefore, to improve performance, we can replace the current strategy of using the AST by a regexp strategy.
This new strategy can lead to false positives (e.g. a string containing import 'foo') but this is not an issue for the hashing function. All that matters is that the hash is consistent and accounts for all files (no false negatives)

* Check for Abs errors