Fix `getSnippetHash` not considering all files Made a stupid mistake in the previous PR: https://github.com/grafana/tanka/pull/759
This fixes it and adds another benchmark test to ensure it doesn't happen again.
I also removed the Github Actions benchmark test, as it's not really useful, anytime we change the tests, we'll get erroneous results which will be annoying.
Instead, I added the benchmark tests to the Drone run, we can compare whenever we want.