Find Importers: Add missing `break` in relative paths logic (#785)
* Find Importers: Add missing `break` in relative paths logic

Jsonnet is weird with relative paths
Given two envs in their own folder. The two following imports in `env1/main.jsonnet will work`: `../env2/main.jsonnet` and `../../env2/main.jsonnet`, so we have to try all of them

The logic does so in a loop. Without a `break`, the logic worked when the match was found on the first case (and correct one) but failed on the second

* Simpler logic + add test

* Add new test