Improved manifest extraction errors Closes https://github.com/grafana/tanka/issues/817
This is something that comes up quite often in our internal environments
When a data structure is not a Kubernetes manifest, the error makes it hard to find out why

This adds a reason, and context for why the object we found isn't a Kubernetes object
In the issue being closed, the error was:

```
>> tk show environments/default
Error: got an error while extracting env `environments/default`: recursion did not resolve in a valid Kubernetes object.  In path `.grafana.deployment` found key `apiVersion` of type `string` instead.
```

It is now:
```
>> tk show environments/default
Error: got an error while extracting env `environments/default`: recursion ended on key "apiVersion" of type string which does not belong to a valid Kubernetes object
instead, it an attribute of the following object:

apiVersion: apps/v1
kind: 3000
metadata:
    name: grafana
spec:
    selector:
        matchLabels:
            name: grafana
    template:
        metadata:
            labels:
                name: grafana
        spec:
            containers:
                - image: grafana/grafana
                  name: grafana
                  ports:
                    - containerPort: 3000
                      name: ui

this object is not a valid Kubernetes object because: attribute "kind" is not a string, it is a float64
```

More verbose, but it actually helps the user find the error