publish-page: Give write permissions explicitly
Dependabot PRs are failing this check, giving a red X in the status that
makes it hard to see whether they passed CI or not. Also, we want this
to work for updates to the yarn stuff.

Then because we do that, we lose the default permissions for other types
too. Specifically, the pr-preview-action needs to comment which is
`pull-requests: write`. So set that too.