Reuse VMs when possible
Whenever ext code and imports are unchanged, VMs have a cache that can be useful for future evaluations.
This PR adds a dumb pool that keeps the latest VM that was used for that same eval path.

I also added a shuffling of envs, so that the likelihood of reusing a VM is higher. When envs are done evaluating, their VM is put back in the pool (with an enriched cache from their eval).
An important thing to note is that VMs cannot be used concurrently, so we need to remove them from the pool when they are used.