Hide the Jsonnet Go implementation behind an interface (#914)
* Hide the Jsonnet Go implementation behind an interface

Tanka currently evals jsonnet using the Go native code.
However, some other implementations have come up in the past years that could be worth using (ex: https://github.com/CertainLach/jrsonnet, which is much faster)

In this PR is the first step: I create an interface where all the jsonnet eval code happens. The Go Jsonnet implementation is now hidden behind this interface.

The setting can either be passed as a global flag or as an env spec attribute to be used when exporting (`spec.exportJsonnetImplementation`)

* PR review changes
`VM` -> `Evaluator`
Add some docstrings
Add some comments

* Select the Jsonnet implementation when constructing the loader

Add the `path` as a struct member of the Go-jsonnet implementation. We
have this available when we construct the loader. This lets us then drop
the same argument from `EvaluateAnonymousSnippet()` in the interface -
as not all implementations need it.

---------

Co-authored-by: Iain Lane <iain.lane@grafana.com>