ci: fix missing tag name in release draft (#1118)
ci: Fix missing tag name in release draft

The `env` context does not contain all environment variables but just
those that were explicitly set by writing to `$GITHUB_ENV` or through
the `env` property of the step/job. Since that is not the case here, the
variable is empty.

This switches over to using `${{ github.ref_name }}`.