feat(ci): run required checks on merge groups
We often get several Dependabot pull requests in a short period. Our
current repository settings - we require a successful build in CI, build
for arm64, and require branches to be up to date - means it can be slow
to manually merge them.

I'd like to switch to using merge queues, then we can disable the
requirement for branches to be up to date before merging, since this
will be handled by the queue. Developers will then be able to review and
enqueue all the PRs, and they will be tested properly against the main
branch.

To be able to do that we need to run the required tests for enqueued
pull requests. That's what we're doing here.