feat: add support for overriding environment label (#824) (#975)
* feat: Add support for overriding environment label

This change adds support for overriding the environment label
using fields on the environment as per #824.

To achieve this, the NameLabel() function that generates this
needed to be lifted to the Environment struct (from Metadata)
to ensure it can access the fields it needs.
Additionally, its signature now returns an error as there are
ways errors could occur during this generation now that should
be surfaced neatly to the user.

* Fix misspellings

---------

Co-authored-by: Julien Duchesne <julien.duchesne@grafana.com>