chore(renovate): use upstream matchers to update bundled binaries
This is Tanka's version of [a `shared-workflows` change][change]. I
noticed that Renovate has [built-in managers][managers] for substituting
versions of things. Here we switch to using those and apply these
consistently in our `Dockerfile` so we're keeping `kubectl`, `jb` (now
downloaded instead of built)`, `helm` and `kustomize` up to date.

Additionally this fixes and enables keeping of Renovate's own version up
to date too.

[change]: https://github.com/grafana/shared-workflows/pull/647
[managers]: https://docs.renovatebot.com/presets-customManagers