chore(docs): bump to tailwind 4.0 (#1327)
This is [a new major version of Tailwind CSS][release-announcement],
which includes a lot of breaking changes. This commit updates the
project to use the new version.

A lot of changes were required to accommodate this. We drop the
`@astrojs/tailwind` plugin since it's not been ported to 4.0 yet. What
that does is bridge between Starlight and Tailwind, and this is
something we can do manually by taking the few definitions that are
needed.

As this implies, the way Tailwind works has changed a fair bit. There's
no Javascript config file any more - it's all done in CSS instead. So
the configuration has had an overhaul.

Some of the classes have changed too. This, and the dropping of the
plugin, meant that the styles broke a bit. We have ported to be more
Tailwind-native to fix this. This means dropping most of the inline css
`<style>` blocks in our components and using the Tailwind native
utililty classes instead. The notable exception is the
`TableOfContentsList` component which is quite custom and so is left for
later.

There are some minor visual differences but overall it should look
broadly the same.

Finally, we've installed some Prettier plugins for Astro and Tailwind,
and ran the code through them.

[release-announcement]: https://blog.tailwindcss.com/tailwindcss-v4