feat(tk-tool-importers): find importers when libraries import from environments (#1663)
This is not a good practice, but nevertheless Tanka does allow it. We're
seeing in our environment that we have situations like:

```jsonnet
{
    foo: 'bar',
}
```

```jsonnet
import ('../../env1/main.jsonnet');
```

```jsonnet
import('lib1/main.libsonnet');
```

Now, in this situation, when `env1` is modified, we also need to ensure
that `env2` is exported. If we don't, then we have drift and there are
changes which aren't applied until some time later.

As far as I can tell, this restriction was a performance optimisation to
avoid searching too many files. I've tested this a fair bit and it
doesn't seem to cause a noticeable performance degradation. One reason
might be that we have all Jsonnet files in memory already, in the
`jsonnetFiles` cache object. So it may not really be very costly to
perform more lookups.